requirements.txt
jinja2==3.1.4
markdown2==2.4.10
pdfkit==1.0.0
python-docx
ell-ai[sqlite]
sqlmodel
pyyaml>=6.0.1
pydantic>=2.0.0
ray>=2.5.0
backoff>=2.2.1
tenacity>=8.2.3
openai>=1.0.0
ray[default]==2.9.0
cloudpickle>=2.2.1
dill>=0.3.8
fastapi==0.109.0
uvicorn==0.24.0.post1
python-multipart==0.0.9
httpx[socks]==0.27.0
socksio==1.0.0
psutil==5.9.8
pytest>=8.2.0
pytest-asyncio>=0.24.0